Refactor: Move stateless code to a new module
Some of the code in chainstate doesn't require a knowledge of the
current tip, and will be used in alternative implementations of
chainstate, like partial_chainstate used in parallel sync.
This commit is move-only.
