+++ crates/floresta-chain/src/pruned_utreexo/mod.rs
@@ -3,6 +3,7 @@ pub mod chain_state;
 pub mod chain_state_builder;
 pub mod chainparams;
 pub mod chainstore;
+pub mod consensus;
 pub mod error;
 pub mod udata;
